Add shorter timeout to graphiql requests
